Town
I made a tile sheet for the town and a building. There is a temp player that has animations as well as WASD movement. The buildings are interactable with E and they load a new scene (what would be the inside of the building but I didn't make those tile sheets yet)